+++ crates/zng-view/src/surface.rs
@@ -315,7 +315,7 @@ impl Surface {
 
         self.push_resize(&mut txn);
 
-        txn.generate_frame(frame.id.get(), render_reasons);
+        txn.generate_frame(frame.id.get(), true, render_reasons);
 
         let frame_scope =
             tracing::trace_span!("<frame>", ?frame.id, capture = ?frame.capture, from_update = false, thread = "<webrender>").entered();
@@ -339,7 +339,7 @@ impl Surface {
         let mut txn = Transaction::new();
         txn.set_root_pipeline(self.pipeline_id);
         self.push_resize(&mut txn);
-        txn.generate_frame(self.frame_id().get(), render_reasons);
+        txn.generate_frame(self.frame_id().get(), true, render_reasons);
 
         let frame_scope = match self.display_list_cache.update(
             &mut DisplayListExtAdapter {