+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+
+# 0.15.4
+
 * Fix interactive carets in rich texts losing pointer capture when crossing over leaf texts.
 * Fix interactive insert caret appearing in non editable text.
 * Update renderer dependencies.